+++ test/Bridges/Bridges.jl
@@ -1,0 +1,19 @@
+include("bridge.jl")
+include("bridgeoptimizer.jl")
+include("singlebridgeoptimizer.jl")
+include("lazybridgeoptimizer.jl")
+@testset "Separate bridges" begin
+    include("flip_sign_bridge.jl")
+    include("vectorizebridge.jl")
+    include("scalarizebridge.jl")
+    include("slackbridge.jl")
+    include("functionize_bridge.jl")
+    include("intervalbridge.jl")
+    include("rsocbridge.jl")
+    include("quadtosocbridge.jl")
+    include("geomeanbridge.jl")
+    include("squarepsdbridge.jl")
+    include("detbridge.jl")
+    include("soctopsdbridge.jl")
+end
+include("external.jl")
