+++ test/Test/intlinear.jl
@@ -1,24 +1,30 @@
-@testset "Integer Linear" begin
-    mock = MOIU.MockOptimizer(Model{Float64}())
-    config = MOIT.TestConfig()
+using Test
+import MathOptInterface
+const MOI = MathOptInterface
+const MOIT = MOI.Test
+const MOIU = MOI.Utilities
 
-    MOIU.set_mock_optimize!(mock,
-        (mock::MOIU.MockOptimizer) -> begin
-            MOI.set(mock, MOI.ObjectiveBound(), 20.0)
-            MOIU.mock_optimize!(mock, [4, 5, 1])
-        end)
-    MOIT.int1test(mock, config)
-    MOIU.set_mock_optimize!(mock,
-        (mock::MOIU.MockOptimizer) -> MOIU.mock_optimize!(mock, [0, 1, 2]),
-        (mock::MOIU.MockOptimizer) -> MOIU.mock_optimize!(mock, [1, 1, 2]),
-        (mock::MOIU.MockOptimizer) -> MOIU.mock_optimize!(mock, [0.0, 0.0, 1.0, 1.0, 0.0, 1.0, 0.0, 0.0, 3.0, 12.0]),
-        (mock::MOIU.MockOptimizer) -> MOIU.mock_optimize!(mock, [0.0, 0.0, 2.0, 2.0, 0.0, 2.0, 0.0, 0.0, 6.0, 24.0]))
-    MOIT.int2test(mock, config)
-    # FIXME [1, 0...] is not the correct optimal solution but it passes the test
-    MOIU.set_mock_optimize!(mock,
-        (mock::MOIU.MockOptimizer) -> MOIU.mock_optimize!(mock, [1.0; zeros(10)]))
-    MOIT.int3test(mock, config)
-    MOIU.set_mock_optimize!(mock,
-        (mock::MOIU.MockOptimizer) -> MOIU.mock_optimize!(mock, [1, 0, 0, 1, 1]))
-    MOIT.knapsacktest(mock, config)
-end
+include("../model.jl")
+
+mock = MOIU.MockOptimizer(Model{Float64}())
+config = MOIT.TestConfig()
+
+MOIU.set_mock_optimize!(mock,
+    (mock::MOIU.MockOptimizer) -> begin
+        MOI.set(mock, MOI.ObjectiveBound(), 20.0)
+        MOIU.mock_optimize!(mock, [4, 5, 1])
+    end)
+MOIT.int1test(mock, config)
+MOIU.set_mock_optimize!(mock,
+    (mock::MOIU.MockOptimizer) -> MOIU.mock_optimize!(mock, [0, 1, 2]),
+    (mock::MOIU.MockOptimizer) -> MOIU.mock_optimize!(mock, [1, 1, 2]),
+    (mock::MOIU.MockOptimizer) -> MOIU.mock_optimize!(mock, [0.0, 0.0, 1.0, 1.0, 0.0, 1.0, 0.0, 0.0, 3.0, 12.0]),
+    (mock::MOIU.MockOptimizer) -> MOIU.mock_optimize!(mock, [0.0, 0.0, 2.0, 2.0, 0.0, 2.0, 0.0, 0.0, 6.0, 24.0]))
+MOIT.int2test(mock, config)
+# FIXME [1, 0...] is not the correct optimal solution but it passes the test
+MOIU.set_mock_optimize!(mock,
+    (mock::MOIU.MockOptimizer) -> MOIU.mock_optimize!(mock, [1.0; zeros(10)]))
+MOIT.int3test(mock, config)
+MOIU.set_mock_optimize!(mock,
+    (mock::MOIU.MockOptimizer) -> MOIU.mock_optimize!(mock, [1, 0, 0, 1, 1]))
+MOIT.knapsacktest(mock, config)
