+++ test/Test/modellike.jl
@@ -1,4 +1,10 @@
-@testset "ModelLike" begin
+using Test
+import MathOptInterface
+const MOI = MathOptInterface
+const MOIT = MOI.Test
+
+@testset "Failed copy" begin
+    include("../dummy.jl")
     model = DummyModelWithAdd()
     MOIT.failcopytestc(model)
     MOIT.failcopytestia(model)
