+++ test/Utilities/cachingoptimizer.jl
@@ -1,4 +1,25 @@
-@MOIU.model ModelForCachingOptimizer (MOI.ZeroOne, MOI.Integer) (MOI.EqualTo, MOI.GreaterThan, MOI.LessThan, MOI.Interval) (MOI.Zeros, MOI.Nonnegatives, MOI.Nonpositives, MOI.SecondOrderCone, MOI.RotatedSecondOrderCone, MOI.GeometricMeanCone, MOI.ExponentialCone, MOI.DualExponentialCone, MOI.PositiveSemidefiniteConeTriangle, MOI.RootDetConeTriangle, MOI.LogDetConeTriangle) () (MOI.SingleVariable,) (MOI.ScalarAffineFunction, MOI.ScalarQuadraticFunction) (MOI.VectorOfVariables,) (MOI.VectorAffineFunction,)
+using Test
+import MathOptInterface
+const MOI = MathOptInterface
+const MOIT = MOI.Test
+const MOIU = MOI.Utilities
+
+include("../model.jl")
+include("../model_for_mock.jl")
+
+MOIU.@model(ModelForCachingOptimizer,
+            (MOI.ZeroOne, MOI.Integer),
+            (MOI.EqualTo, MOI.GreaterThan, MOI.LessThan, MOI.Interval),
+            (MOI.Zeros, MOI.Nonnegatives, MOI.Nonpositives, MOI.SecondOrderCone,
+             MOI.RotatedSecondOrderCone, MOI.GeometricMeanCone,
+             MOI.ExponentialCone, MOI.DualExponentialCone,
+             MOI.PositiveSemidefiniteConeTriangle, MOI.RootDetConeTriangle,
+             MOI.LogDetConeTriangle),
+            (),
+            (MOI.SingleVariable,),
+            (MOI.ScalarAffineFunction, MOI.ScalarQuadraticFunction),
+            (MOI.VectorOfVariables,),
+            (MOI.VectorAffineFunction,))
 
 @testset "Test default attributes" begin
     # Without an optimizer attached (i.e., `MOI.state(model) == NO_OPTIMIZER`) we
@@ -241,7 +262,7 @@ end
 
 end
 
-@testset "CachingOptimizer constructor with optimizer" begin
+@testset "Constructor with optimizer" begin
     @testset "Empty model and optimizer" begin
         s = MOIU.MockOptimizer(ModelForMock{Float64}(), supports_names=false)
         model = ModelForCachingOptimizer{Float64}()
