+++ test/Utilities/constraints.jl
@@ -1,4 +1,10 @@
-@testset "Scalar constraints" begin
+using Test
+import MathOptInterface
+const MOI = MathOptInterface
+
+include("../model.jl")
+
+@testset "Scalar" begin
     model = Model{Float64}()
     x = MOI.add_variable(model)
     @testset "SingleVariable" begin
