+++ test/runtests.jl
@@ -21,45 +21,9 @@ include("dummy.jl")
     include("attributes.jl")
 end
 
-# Needed by test spread over several files, defining it here make it easier to comment out tests
-# Model supporting every MOI functions and sets
-MOIU.@model(Model,
-            (MOI.ZeroOne, MOI.Integer),
-            (MOI.EqualTo, MOI.GreaterThan, MOI.LessThan, MOI.Interval,
-             MOI.Semicontinuous, MOI.Semiinteger),
-            (MOI.Reals, MOI.Zeros, MOI.Nonnegatives, MOI.Nonpositives,
-             MOI.SecondOrderCone, MOI.RotatedSecondOrderCone,
-             MOI.GeometricMeanCone, MOI.ExponentialCone, MOI.DualExponentialCone,
-             MOI.PositiveSemidefiniteConeTriangle, MOI.PositiveSemidefiniteConeSquare,
-             MOI.RootDetConeTriangle, MOI.RootDetConeSquare, MOI.LogDetConeTriangle,
-             MOI.LogDetConeSquare),
-            (MOI.PowerCone, MOI.DualPowerCone, MOI.SOS1, MOI.SOS2),
-            (MOI.SingleVariable,),
-            (MOI.ScalarAffineFunction, MOI.ScalarQuadraticFunction),
-            (MOI.VectorOfVariables,),
-            (MOI.VectorAffineFunction, MOI.VectorQuadraticFunction))
-
-# Model supporting only SecondOrderCone as non-LP cone.
-MOIU.@model(ModelForMock, (MOI.ZeroOne, MOI.Integer),
-            (MOI.EqualTo, MOI.GreaterThan, MOI.LessThan, MOI.Interval),
-            (MOI.Zeros, MOI.Nonnegatives, MOI.Nonpositives, MOI.SecondOrderCone),
-            (),
-            (MOI.SingleVariable,),
-            (MOI.ScalarAffineFunction, MOI.ScalarQuadraticFunction),
-            (MOI.VectorOfVariables,),
-            (MOI.VectorAffineFunction,))
-
 # Utilities submodule tests
 @testset "MOI.Utilities" begin
-    include("Utilities/functions.jl")
-    include("Utilities/sets.jl")
-    include("Utilities/constraints.jl")
-    include("Utilities/model.jl")
-    include("Utilities/universalfallback.jl")
-    include("Utilities/parser.jl")
-    include("Utilities/mockoptimizer.jl")
-    include("Utilities/cachingoptimizer.jl")
-    include("Utilities/copy.jl")
+    include("Utilities/Utilities.jl")
 end
 
 # Test submodule tests
